Await the WriteAsync calls
While this may not be strictly necessary in the BAD or GOOD examples,
omitting the await may be misinterpreted as a required condition for the BAD case.
To avoid people rationalizing leaving their code alone by thinking "well, I always await the WriteAsync call, so I'm good",
I'm adding the await call to the sample to show that even that doesn't mitigate the problem.